Match KnockDownPlank::InitResources byte-identical (mwccarm 2004/b56)
ov015 _ZN14KnockDownPlank13InitResourcesEv @ 0x02111960 size 0x208.
Levers: opt_propagation off, u64-mask angle increment, named zero for
the two rotated-offset Vector3 stores. linkcheck VERIFIED.
