+++ config/arm9/overlays/ov015/delinks.txt
@@ -72,6 +72,10 @@ src/_ZN14KnockDownPlank8BehaviorEv.cpp:
     complete
     .text start:0x02111720 end:0x02111960
 
+src/_ZN14KnockDownPlank13InitResourcesEv.cpp:
+    complete
+    .text start:0x02111960 end:0x02111b68
+
 src/KnockDownPlank_Spawn.c:
     complete
     .text start:0x02111b68 end:0x02111ba0
