+++ config/match_attempts.jsonl
@@ -1588,3 +1588,8 @@
 {"schemaVersion":1,"functionId":"ov006:0x020ee994","id":"ov006:0x020ee994","attemptId":"37ffdee582ef4911a8acc5935974fa3f","parentAttemptId":"36beb73083604118a53a387d850e5599","module":"ov006","addr":34531732,"name":"func_ov006_020ee994","status":"no_progress","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":999,"prevBestDivergences":6,"improvedNearMiss":false,"srcPath":"scratch/func_ov006_020ee994_duff.c","note":"duff/switch second entry remats but adds jump-table (+6-7 insn). if(0)+dead_code off DCE loses irreducibility. named temps/ten-live/C++ member access fold to same tip IR (mul r0). interleave *lp schedules ldr-last between mul/asr but last in r1 not r0. wall reconfirm: notes/mwccarm-codegen.md const-hoist.","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"36beb73083604118a53a387d850e5599","divergences":6},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov006:0x020ee994","id":"ov006:0x020ee994","attemptId":"4337326d12fe455a97068925871cb39b","parentAttemptId":"36beb73083604118a53a387d850e5599","module":"ov006","addr":34531732,"name":"func_ov006_020ee994","status":"no_progress","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":999,"prevBestDivergences":6,"improvedNearMiss":false,"srcPath":"scratch/func_ov006_020ee994.c","note":"clean rewrite from m2c (Ghidra unavailable). computed-goto unsupported. launder/volatile/O0-mix/pragmas still hoist. dScMgBase C++ header not C-legal.","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"36beb73083604118a53a387d850e5599","divergences":6},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov006:0x020ee994","id":"ov006:0x020ee994","attemptId":"16e4efbd49ec4ad78a28095c4dc9d137","parentAttemptId":"36beb73083604118a53a387d850e5599","module":"ov006","addr":34531732,"name":"func_ov006_020ee994","status":"no_progress","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":999,"prevBestDivergences":6,"improvedNearMiss":false,"note":"permuter 465it on do-while: best permuter scores 155-435 pulled Random out of loop (semantically broken infinite reroll). Faithful perms stay hoisted size 0x164. long-long temp + tip still mul r0 / extra cmp+blo. wall holds.","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"36beb73083604118a53a387d850e5599","divergences":6},"sessionScope":"focused","batchSize":1}
+{"schemaVersion":1,"functionId":"ov015:0x02111960","id":"ov015:0x02111960","attemptId":"5bf53f607a7541a69180dcd01e187377","parentAttemptId":null,"module":"ov015","addr":34675040,"name":"_ZN14KnockDownPlank13InitResourcesEv","status":"near_miss","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":40,"prevBestDivergences":999,"improvedNearMiss":true,"note":"First compile: shared headers, u64-mask angle, four Vector3s declared d-c-b-a so stack slots reversed vs ROM.","usedNearMissDraft":false,"usedGhidraDraft":false,"base":{"kind":"scratch"},"sessionScope":"focused","batchSize":1}
+{"schemaVersion":1,"functionId":"ov015:0x02111960","id":"ov015:0x02111960","attemptId":"6ba7374529754e358150110f77401b80","parentAttemptId":"5bf53f607a7541a69180dcd01e187377","module":"ov015","addr":34675040,"name":"_ZN14KnockDownPlank13InitResourcesEv","status":"near_miss","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":19,"prevBestDivergences":40,"improvedNearMiss":true,"note":"Declare Vector3 a,b,c,d so slots match ROM. Residual: angle RMW r2/r1 vs r3/r2 and zero remat in r0.","usedNearMissDraft":false,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"5bf53f607a7541a69180dcd01e187377","divergences":40},"sessionScope":"focused","batchSize":1}
+{"schemaVersion":1,"functionId":"ov015:0x02111960","id":"ov015:0x02111960","attemptId":"09be0c40ec714038a323bfad7836f245","parentAttemptId":"6ba7374529754e358150110f77401b80","module":"ov015","addr":34675040,"name":"_ZN14KnockDownPlank13InitResourcesEv","status":"no_progress","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":19,"prevBestDivergences":19,"improvedNearMiss":false,"note":"Named pang/ang/zval/zero temps optimized away; same 19-word residual.","usedNearMissDraft":false,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"6ba7374529754e358150110f77401b80","divergences":19},"sessionScope":"focused","batchSize":1}
+{"schemaVersion":1,"functionId":"ov015:0x02111960","id":"ov015:0x02111960","attemptId":"ef8f9685fd984c9392b1a367fc25295b","parentAttemptId":"6ba7374529754e358150110f77401b80","module":"ov015","addr":34675040,"name":"_ZN14KnockDownPlank13InitResourcesEv","status":"near_miss","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":12,"prevBestDivergences":19,"improvedNearMiss":true,"note":"opt_propagation off: angle base r3 and second vec-init match. Residual first-block hoist of 0xfa000 vs ldrsh.","usedNearMissDraft":false,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"6ba7374529754e358150110f77401b80","divergences":19},"sessionScope":"focused","batchSize":1}
+{"schemaVersion":1,"functionId":"ov015:0x02111960","id":"ov015:0x02111960","attemptId":"ddaf0bbbdc9248ed86b4aa24ac55079d","parentAttemptId":"ef8f9685fd984c9392b1a367fc25295b","module":"ov015","addr":34675040,"name":"_ZN14KnockDownPlank13InitResourcesEv","status":"matched","kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build","author":"lunavyqo","divergences":0,"prevBestDivergences":12,"improvedNearMiss":true,"srcPath":"src/_ZN14KnockDownPlank13InitResourcesEv.cpp","note":"opt_propagation off + named zero + a.z=-0xfa000 after u64-mask angle increment. linkcheck VERIFIED.","usedNearMissDraft":false,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"ef8f9685fd984c9392b1a367fc25295b","divergences":12},"sessionScope":"focused","batchSize":1}
