+++ config/match_provenance.jsonl
@@ -383,3 +383,4 @@
 {"id":"arm9:0x0204504c","module":"arm9","addr":33837132,"name":"_ZN15ModelComponents21UpdateVertsUsingBonesEv","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"tangosdev","srcPath":"src/_ZN15ModelComponents21UpdateVertsUsingBonesEv.c"}
 {"id":"arm9:0x02045394","module":"arm9","addr":33837972,"name":"_ZN15ModelComponents11UpdateBonesEP8BCA_Filei","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"tangosdev","srcPath":"src/_ZN15ModelComponents11UpdateBonesEP8BCA_Filei.cpp"}
 {"id":"arm9:0x0206cd44","module":"arm9","addr":34000196,"name":"_ZN4cstd8__assertEPKcS1_S1_i","matchProvenance":{"kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/_ZN4cstd8__assertEPKcS1_S1_i.cpp"}
+{"id":"ov015:0x02111960","module":"ov015","addr":34675040,"name":"_ZN14KnockDownPlank13InitResourcesEv","matchProvenance":{"kind":"ai","model":"grok-4.6","reasoning":"xhigh","harness":"grok-build"},"author":"lunavyqo","srcPath":"src/_ZN14KnockDownPlank13InitResourcesEv.cpp"}
